Projet_Machine_Learning/requirements.txt
altair==5.1.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.7
executing==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.33
idna==3.4
importlib-metadata==6.8.0
imutils==0.5.4
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
psycopg2-binary==2.9.7
kiwisolver==1.4.5
libclang==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpld3==0.5.9
ncat==0.1.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
plotly==5.16.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==13.0.0
pyasn1==0.5.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
pyspark==3.4.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
tenacity==8.2.3
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.21.2
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
